backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
celery[redis]>=5.3.0
redis>=5.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
httpx>=0.26.0
pandas>=2.0.0
supabase>=2.0.0
PyJWT>=2.8.0
email-validator>=2.0.0
